+++ perf/client_consistency_test.go
