+++ perf/go.mod
@@ -1,8 +1,9 @@
-module github.com/grafana/nanogit/tests/performance
+module github.com/grafana/nanogit/perf
 
 go 1.24.4
 
 require (
+	github.com/go-git/go-billy/v5 v5.5.0
 	github.com/go-git/go-git/v5 v5.12.0
 	github.com/grafana/nanogit v0.0.0-00010101000000-000000000000
 	github.com/stretchr/testify v1.10.0
@@ -29,7 +30,6 @@ require (
 	github.com/emirpasic/gods v1.18.1 // indirect
 	github.com/felixge/httpsnoop v1.0.4 // indirect
 	github.com/go-git/gcfg v1.5.1-0.20230307220236-3a3c6141e376 // indirect
-	github.com/go-git/go-billy/v5 v5.5.0 // indirect
 	github.com/go-logr/logr v1.4.2 // indirect
 	github.com/go-logr/stdr v1.2.2 // indirect
 	github.com/go-ole/go-ole v1.2.6 // indirect
@@ -78,4 +78,4 @@ require (
 	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
 
-replace github.com/grafana/nanogit => ../..
+replace github.com/grafana/nanogit => ..
